ported to samsung i9000 by modifying aries-rfkill.c instead of herring-rfkill.c
cpuidle: idle2: v0.210 standalone patch
Contains squashed commit of: v0.201, v0.202, v0.210 patches.

cpuidle: idle2: v0.210
Implement DEEP-IDLE TOP ON state for use when bluetooth or GPS is active
Many micro-optimisations of functions and variables to speed up the hot paths
Changed some debug printk to KERN_DEBUG.

cpuidle: idle2: v0.202
Fix issue where IDLE2 mode would not be entered after antutu had been ran.
Issue was related to power gating checks from deep-idle which I implemented
in v0.200. These checks are not needed, and cause this issue due to the
S5PV210_PD_CAM power domain not being gated after antutu and some other
benchmarks have been ran.

v0.201 is a complete rewrite of the majority of the idle2 codebase.
2 idle states have now been fully implemented and are available at all times.
The states are IDLE and IDLE2, which is DEEP-IDLE TOP OFF.
DEEP-IDLE TOP ON is not implemented, as it has no benefit.

The hot code paths are now more optimised and faster.
Workqueues are used to inform idle2 of events from other subsystems,
such as audio, usb connected and bluetooth.
PM notifiers are used to notify idle2 of a suspend, although it should
never get into a situation where it is working at the same time as the
system is suspending.

Add code to disable cpufreq and lock the CPU at 800MHz when the conditions
are right to enter IDLE2. This appears to save a lot more power.
Add earlysuspend hook back in to ensure the above doesn't occur whilst the
screen is on.

Adjust exit latency to 400 as that is the max specified in the TRM. Also
reduce residency time to 2000 as that seems like a more reasonable figure.
Remove s5p_idle2 function from asm and use generic flush_cache_all()
and cpu_do_idle() calls instead. The only asm left now is the register
save and restore code.

The IDLE2 mode will only be activated when audio has been playing back
for more than 30 seconds, the device is not connected to USB, bluetooth is
not active (it can be enabled, but not be actively working) various clocks
are gated (in practice, this means that the screen has to be off), certain
subsystems are not active and there are no interrupts pending.

The early suspend and late resume hooks have been removed as they are not
required, as have the suspend and resume hooks.

Many of the idle2 specific functions have been split into idle2.h to keep
them separate from the stock code.

IDLE2 can be disabled with the following command:
echo 1 > /sys/module/cpuidle/parameters/idle2_disabled

Stats can be obtained from the standard cpuidle interface, which, for
IDLE2 is located at:
/sys/devices/system/cpu/cpu0/cpuidle/state1