idle2: v0.390 Hopefully fix bluetooth streaming / tethering issue by rewriting bluetooth activity detection code. Also rewrite the UART detection code and remove the complicated idle2_cancel_topon_work_fn() as it's all rewritten to be much simpler by separating out the two requests. The UART code has been simplified the delayed work has been removed and it is a simple toggle now to set a flag.
The bluetooth detection has been rewritten as a oneshot function which is
fired every time the bluetooth host wake IRQ handler is invoked and disables
TOP block off for 10 seconds each time it is fired.
If bluetooth streaming is active, this will be fired regularly, causing the
SoC to stay in TOP block on mode and allow the music to continue.

Flush caches after waking to ensure they are clear

IDLE2 will now only be active when the screen is off. This should work
around any reported performance issues. The actual benefit of IDLE2 whilst
the screen is on is pretty much nil anyway.

Optimise code for TOP block off fast path.

Use cancel_delayed_work() instead of cancel_delayed_work_sync() as we are
atomic and cancel_delayed_work_sync() might_sleep() which causes a BUG().

Net deletion of 56 lines of code, so some of the bloat is going. :p